further isolate animation code, add dancing to SM
